+++ Yvand.LDAPCPSE.Tests/UnitTestsHelper.cs
@@ -16,42 +16,36 @@ namespace Yvand.LdapClaimsProvider.Tests
     [SetUpFixture]
     public class UnitTestsHelper
     {
-        public static readonly LDAPCPSE ClaimsProvider = new LDAPCPSE(TestContext.Parameters["ClaimsProviderName"]);
-        public static SPTrustedLoginProvider SPTrust => SPSecurityTokenServiceManager.Local.TrustedLoginProviders.FirstOrDefault(x => String.Equals(x.ClaimProviderName, TestContext.Parameters["ClaimsProviderName"], StringComparison.InvariantCultureIgnoreCase));
+        public static readonly string ClaimsProviderName = TestContext.Parameters["ClaimsProviderName"];
+        public static readonly LDAPCPSE ClaimsProvider = new LDAPCPSE(ClaimsProviderName);
+        public static SPTrustedLoginProvider SPTrust => Utils.GetSPTrustAssociatedWithClaimsProvider(ClaimsProviderName);
         public static Uri TestSiteCollUri;
-        public static string TestSiteRelativePath => $"/sites/{TestContext.Parameters["TestSiteCollectionName"]}";
+        public static string TestSiteRelativePath => $"/sites/{ClaimsProviderName}.UnitTests";
         public const int MaxTime = 50000;
+        public const int TestRepeatCount = 1;
         public static string FarmAdmin => TestContext.Parameters["FarmAdmin"];
         public static string DomainFqdn => TestContext.Parameters["DomainFqdn"];
-        public static string Domain => TestContext.Parameters["Domain"];
-#if DEBUG
-        public const int TestRepeatCount = 1;
-#else
-    public const int TestRepeatCount = 20;
-#endif
 
         public static string RandomClaimType => "http://schemas.yvand.net/ws/claims/random";
         public static string RandomClaimValue => "IDoNotExist";
         public static string RandomDirectoryObjectClass => "randomClass";
         public static string RandomDirectoryObjectAttribute => "randomAttribute";
 
-        public static string ValidGroupClaimType => TestContext.Parameters["ValidGroupClaimType"];
-        public static string ValidGroupName => TestContext.Parameters["ValidGroupName"];
-        public const string ValidGroupSid = "S-1-5-21-2647467245-1611586658-188888215-11601"; //=> TestContext.Parameters["ValidTrustedGroupName"]; // CONTOSO\group1
-        public const string ValidUserSid = "S-1-5-21-2647467245-1611586658-188888215-107206"; // => TestContext.Parameters["ValidUserSid"]; // CONTOSO\testLdapcpseUser_001
+        public static string GroupsClaimType => TestContext.Parameters["GroupsClaimType"];
 
-        public static string AzureTenantsJsonFile => TestContext.Parameters["LdapConnections"];
+        public static string LdapConnectionsJsonFile => TestContext.Parameters["LdapConnectionsJsonFile"];
         public static string DataFile_AllAccounts_Search => TestContext.Parameters["DataFile_AllAccounts_Search"];
         public static string DataFile_AllAccounts_Validate => TestContext.Parameters["DataFile_AllAccounts_Validate"];
+        public static string DataFile_TestUsers => TestContext.Parameters["DataFile_TestUsers"];
+        public static string DataFile_TestGroups => TestContext.Parameters["DataFile_TestGroups"];
         static TextWriterTraceListener Logger { get; set; }
         public static LdapProviderConfiguration PersistedConfiguration;
         private static ILdapProviderSettings OriginalSettings;
 
-
         [OneTimeSetUp]
         public static void InitializeSiteCollection()
         {
-            Logger = new TextWriterTraceListener(TestContext.Parameters["TestLogFileName"]);
+            Logger = new TextWriterTraceListener($"{ClaimsProviderName}IntegrationTests.log");
             Trace.Listeners.Add(Logger);
             Trace.AutoFlush = true;
             Trace.TraceInformation($"{DateTime.Now.ToString("s")} [SETUP] Start integration tests of {ClaimsProvider.Name} {FileVersionInfo.GetVersionInfo(Assembly.GetAssembly(typeof(LDAPCPSE)).Location).FileVersion}.");
@@ -82,7 +76,7 @@ namespace Yvand.LdapClaimsProvider.Tests
 
 #if DEBUG
             TestSiteCollUri = new Uri($"http://spsites{TestSiteRelativePath}");
-            return; // Uncommented when debugging from unit tests
+            //return; // Uncommented when debugging from unit tests
 #endif
 
             var service = SPFarm.Local.Services.GetValue<SPWebService>(String.Empty);
@@ -110,8 +104,9 @@ namespace Yvand.LdapClaimsProvider.Tests
             Uri waRootAuthority = wa.AlternateUrls[0].Uri;
             TestSiteCollUri = new Uri($"{waRootAuthority.GetLeftPart(UriPartial.Authority)}{TestSiteRelativePath}");
             SPClaimProviderManager claimMgr = SPClaimProviderManager.Local;
-            string encodedGroupClaim = claimMgr.EncodeClaim(new SPClaim(ValidGroupClaimType, ValidGroupName, ClaimValueTypes.String, SPOriginalIssuers.Format(SPOriginalIssuerType.TrustedProvider, UnitTestsHelper.SPTrust.Name)));
-            SPUserInfo groupInfo = new SPUserInfo { LoginName = encodedGroupClaim, Name = ValidGroupName };
+            string trustedGroupName = TestEntitySourceManager.GetOneGroup().AccountNameFqdn;
+            string encodedGroupClaim = claimMgr.EncodeClaim(new SPClaim(GroupsClaimType, trustedGroupName, ClaimValueTypes.String, SPOriginalIssuers.Format(SPOriginalIssuerType.TrustedProvider, UnitTestsHelper.SPTrust.Name)));
+            SPUserInfo groupInfo = new SPUserInfo { LoginName = encodedGroupClaim, Name = trustedGroupName };
 
             FileVersionInfo spAssemblyVersion = FileVersionInfo.GetVersionInfo(Assembly.GetAssembly(typeof(SPSite)).Location);
             string spSiteTemplate = "STS#3"; // modern team site template
@@ -179,15 +174,6 @@ namespace Yvand.LdapClaimsProvider.Tests
         }
     }
 
-    //public class SearchEntityDataSourceCollection : IEnumerable
-    //{
-    //    public IEnumerator GetEnumerator()
-    //    {
-    //        yield return new[] { "AADGroup1", "1", "5b0f6c56-c87f-44c3-9354-56cba03da433" };
-    //        yield return new[] { "AADGroupTes", "1", "99abdc91-e6e0-475c-a0ba-5014f91de853" };
-    //    }
-    //}
-
     public enum ResultEntityType
     {
         None,
@@ -196,79 +182,232 @@ namespace Yvand.LdapClaimsProvider.Tests
         Group,
     }
 
-    public enum EntityDataSourceType
+    public abstract class TestEntity : ICloneable
     {
-        AllAccounts,
-        UPNB2BGuestAccounts
+        public object Clone()
+        {
+            return this.MemberwiseClone();
+        }
+
+        public abstract void SetEntityFromDataSourceRow(Row row);
     }
 
-    public class SearchEntityData
+    public class TestUser : TestEntity
+    {
+        public string UserPrincipalName;
+        public string SamAccountName;
+        public string Mail;
+        public string GivenName;
+        public string DisplayName;
+        public string DistinguishedName;
+        public string SID;
+        public bool IsMemberOfAllGroups;
+        public bool IsMemberOfNestedGroups;
+
+        public override void SetEntityFromDataSourceRow(Row row)
+        {
+            UserPrincipalName = row["userPrincipalName"];
+            SamAccountName = row["SamAccountName"];
+            Mail = row["mail"];
+            GivenName = row["givenName"];
+            DisplayName = row["displayName"];
+            DistinguishedName = row["DistinguishedName"];
+            SID = row["SID"];
+            IsMemberOfAllGroups = Convert.ToBoolean(row["IsMemberOfAllGroups"]);
+            IsMemberOfNestedGroups = Convert.ToBoolean(row["IsMemberOfNestedGroups"]);
+        }
+    }
+
+    public class TestGroup : TestEntity
+    {
+        public string SamAccountName;
+        public string DistinguishedName;
+        public string SID;
+        public bool EveryoneIsMember;
+        public string AccountNameFqdn;
+        public bool IsNestedGroup;
+        public bool NestedGroupsAreMembers;
+
+        public override void SetEntityFromDataSourceRow(Row row)
+        {
+            SamAccountName = row["SamAccountName"];
+            DistinguishedName = row["DistinguishedName"];
+            SID = row["SID"];
+            EveryoneIsMember = Convert.ToBoolean(row["EveryoneIsMember"]);
+            AccountNameFqdn = row["AccountNameFqdn"];
+            IsNestedGroup = Convert.ToBoolean(row["IsNestedGroup"]);
+            NestedGroupsAreMembers = Convert.ToBoolean(row["NestedGroupsAreMembers"]);
+        }
+    }
+
+    public class SearchEntityScenario : TestEntity
     {
         public string Input;
         public int SearchResultCount;
         public string SearchResultSingleEntityClaimValue;
         public ResultEntityType SearchResultEntityTypes;
         public bool ExactMatch;
-    }
 
-    public class SearchEntityDataSource
-    {
-        public static IEnumerable<TestCaseData> GetTestData(EntityDataSourceType entityDataSourceType)
+        public override void SetEntityFromDataSourceRow(Row row)
         {
-            string csvPath = UnitTestsHelper.DataFile_AllAccounts_Search;
-            DataTable dt = DataTable.New.ReadCsv(csvPath);
-            foreach (Row row in dt.Rows)
-            {
-                var registrationData = new SearchEntityData();
-                registrationData.Input = row["Input"];
-                registrationData.SearchResultCount = Convert.ToInt32(row["SearchResultCount"]);
-                registrationData.SearchResultSingleEntityClaimValue = row["SearchResultSingleEntityClaimValue"];
-                registrationData.SearchResultEntityTypes = (ResultEntityType)Enum.Parse(typeof(ResultEntityType), row["SearchResultEntityTypes"]);
-                registrationData.ExactMatch = Convert.ToBoolean(row["ExactMatch"]);
-                yield return new TestCaseData(new object[] { registrationData });
-            }
-        }
-
-        //public class ReadCSV
-        //{
-        //    public void GetValue()
-        //    {
-        //        TextReader tr1 = new StreamReader(@"c:\pathtofile\filename", true);
-
-        //        var Data = tr1.ReadToEnd().Split('\n')
-        //        .Where(l => l.Length > 0)  //nonempty strings
-        //        .Skip(1)               // skip header 
-        //        .Select(s => s.Trim())   // delete whitespace
-        //        .Select(l => l.Split(',')) // get arrays of values
-        //        .Select(l => new { Field1 = l[0], Field2 = l[1], Field3 = l[2] });
-        //    }
-        //}
-    }
-
-    public class ValidateEntityDataSource
-    {
-        public static IEnumerable<TestCaseData> GetTestData(EntityDataSourceType entityDataSourceType)
-        {
-            string csvPath = UnitTestsHelper.DataFile_AllAccounts_Validate;
-            DataTable dt = DataTable.New.ReadCsv(csvPath);
-
-            foreach (Row row in dt.Rows)
-            {
-                var registrationData = new ValidateEntityData();
-                registrationData.ClaimValue = row["ClaimValue"];
-                registrationData.ShouldValidate = Convert.ToBoolean(row["ShouldValidate"]);
-                registrationData.IsMemberOfTrustedGroup = Convert.ToBoolean(row["IsMemberOfTrustedGroup"]);
-                registrationData.EntityType = (ResultEntityType)Enum.Parse(typeof(ResultEntityType), row["EntityType"]);
-                yield return new TestCaseData(new object[] { registrationData });
-            }
+            Input = row["Input"];
+            SearchResultCount = Convert.ToInt32(row["SearchResultCount"]);
+            SearchResultSingleEntityClaimValue = row["SearchResultSingleEntityClaimValue"];
+            SearchResultEntityTypes = (ResultEntityType)Enum.Parse(typeof(ResultEntityType), row["SearchResultEntityTypes"]);
+            ExactMatch = Convert.ToBoolean(row["ExactMatch"]);
         }
     }
 
-    public class ValidateEntityData
+    public class ValidateEntityScenario : TestEntity
     {
         public string ClaimValue;
         public bool ShouldValidate;
         public bool IsMemberOfTrustedGroup;
         public ResultEntityType EntityType;
+
+        public override void SetEntityFromDataSourceRow(Row row)
+        {
+            ClaimValue = row["ClaimValue"];
+            ShouldValidate = Convert.ToBoolean(row["ShouldValidate"]);
+            IsMemberOfTrustedGroup = Convert.ToBoolean(row["IsMemberOfTrustedGroup"]);
+            EntityType = (ResultEntityType)Enum.Parse(typeof(ResultEntityType), row["EntityType"]);
+        }
+    }
+
+    public class TestEntitySource<T> where T : TestEntity, new()
+    {
+        private object _LockInitEntitiesList = new object();
+        private bool EntitiesReady = false;
+        private List<T> _Entities;
+        public List<T> Entities
+        {
+            get
+            {
+                if (EntitiesReady) { return _Entities; }
+                lock (_LockInitEntitiesList)
+                {
+                    if (EntitiesReady) { return _Entities; }
+                    _Entities = new List<T>();
+                    foreach (T entity in ReadDataSource())
+                    {
+                        _Entities.Add(entity);
+                    }
+                    EntitiesReady = true;
+                    Trace.TraceInformation($"{DateTime.Now:s} [{typeof(T).Name}] Initialized List of {nameof(Entities)} with {Entities.Count} items.");
+                    return _Entities;
+                }
+            }
+        }
+
+        private Random RandomNumber = new Random();
+        private string DataSourceFilePath;
+
+        public TestEntitySource(string dataSourceFilePath)
+        {
+            DataSourceFilePath = dataSourceFilePath;
+        }
+
+        private IEnumerable<T> ReadDataSource()
+        {
+            DataTable dt = DataTable.New.ReadCsv(DataSourceFilePath);
+            foreach (Row row in dt.Rows)
+            {
+                T entity = new T();
+                entity.SetEntityFromDataSourceRow(row);
+                yield return entity;
+            }
+        }
+
+        public IEnumerable<T> GetSomeEntities(int count, Func<T, bool> filter = null)
+        {
+            IEnumerable<T> entitiesFiltered = Entities.Where(filter ?? (x => true));
+            int entitiesFilteredCount = entitiesFiltered.Count();
+            if (count > entitiesFilteredCount) { count = entitiesFilteredCount; }
+            for (int i = 0; i < count; i++)
+            {
+                int randomIdx = RandomNumber.Next(0, entitiesFilteredCount);
+                yield return entitiesFiltered.ElementAt(randomIdx).Clone() as T;
+            }
+        }
+    }
+
+    public class TestEntitySourceManager
+    {
+#if DEBUG
+
+        private static TestEntitySource<SearchEntityScenario> SearchTestsSource = new TestEntitySource<SearchEntityScenario>(UnitTestsHelper.DataFile_AllAccounts_Search);
+        public static List<SearchEntityScenario> AllSearchEntities
+        {
+            get => SearchTestsSource.Entities;
+        }
+        private static TestEntitySource<ValidateEntityScenario> ValidationTestsSource = new TestEntitySource<ValidateEntityScenario>(UnitTestsHelper.DataFile_AllAccounts_Validate);
+        public static List<ValidateEntityScenario> AllValidationEntities
+        {
+            get => ValidationTestsSource.Entities;
+        }
+#endif
+        private static TestEntitySource<TestUser> TestUsersSource = new TestEntitySource<TestUser>(UnitTestsHelper.DataFile_TestUsers);
+        public static List<TestUser> AllTestUsers
+        {
+            get => TestUsersSource.Entities;
+        }
+        private static TestEntitySource<TestGroup> TestGroupsSource = new TestEntitySource<TestGroup>(UnitTestsHelper.DataFile_TestGroups);
+        public static List<TestGroup> AllTestGroups
+        {
+            get => TestGroupsSource.Entities;
+        }
+        public const int MaxNumberOfUsersToTest = 100;
+        public const int MaxNumberOfGroupsToTest = 50;
+
+        public static IEnumerable<TestUser> GetSomeUsers(int count)
+        {
+            return TestUsersSource.GetSomeEntities(count, null);
+        }
+
+        public static IEnumerable<TestUser> GetUsersMembersOfAllGroups()
+        {
+            Func<TestUser, bool> filter = x => x.IsMemberOfAllGroups == true;
+            return TestUsersSource.GetSomeEntities(Int16.MaxValue, filter);
+        }
+
+        public static IEnumerable<TestUser> GetUsersMembersOfNestedGroups()
+        {
+            Func<TestUser, bool> filter = x => x.IsMemberOfNestedGroups == true;
+            return TestUsersSource.GetSomeEntities(Int16.MaxValue, filter);
+        }
+
+        public static TestUser FindUser(string upnPrefix)
+        {
+            Func<TestUser, bool> filter = x => x.UserPrincipalName.StartsWith(upnPrefix);
+            return TestUsersSource.GetSomeEntities(1, filter).First();
+        }
+
+        public static IEnumerable<TestGroup> GetSomeGroups(int count)
+        {
+            return TestGroupsSource.GetSomeEntities(count);
+        }
+
+        public static IEnumerable<TestGroup> GetNestedGroups(int count)
+        {
+            Func<TestGroup, bool> filter = x => x.IsNestedGroup == true;
+            return TestGroupsSource.GetSomeEntities(count, filter);
+        }
+
+        public static IEnumerable<TestGroup> GetGroupsWithNestedGroupsAsMembers(int count)
+        {
+            Func<TestGroup, bool> filter = x => x.NestedGroupsAreMembers == true;
+            return TestGroupsSource.GetSomeEntities(count, filter);
+        }
+
+        public static TestGroup GetOneGroup()
+        {
+            return TestGroupsSource.GetSomeEntities(1, null).First();
+        }
+
+        public static TestGroup FindGroup(string groupName)
+        {
+            Func<TestGroup, bool> filter = x => String.Equals(x.SamAccountName, groupName, StringComparison.OrdinalIgnoreCase);
+            return TestGroupsSource.GetSomeEntities(1, filter).First();
+        }
     }
 }