+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change log for LDAPCP
 
+## Unreleased
+
+* Fix error when creating the configuration if the trust uses an identifier claim type that is not well-known by LDAPCP - https://github.com/Yvand/LDAPCP/issues/221
+* Improve tests
+
 ## LDAPCP Second Edition v18.0.20240513.3 - Published in May 13, 2024
 
 * Fix error when creating the configuration, due to case-sensitive test in the claim types - https://github.com/Yvand/LDAPCP/issues/204